@@ -5,13 +5,13 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.integration.aggregator.MessageGroupProcessor;
 import org.springframework.integration.aggregator.ReleaseStrategy;
 import org.springframework.integration.config.EnableIntegration;
 import org.springframework.integration.dsl.IntegrationFlow;
-import org.springframework.integration.dsl.IntegrationFlows;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.support.MessageBuilder;
@@ -28,15 +28,17 @@ public class IntegrationConfiguration {
 
     private final Function<String, String[]> splitWordsFunction = sentence -> sentence.split(" ");
     private final Function<List<String>, Map<String, Long>> convertArrayListToCountMap = list -> list.stream()
-      .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
+        .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
     private final Function<String, String> toLowerCase = String::toLowerCase;
     private final MessageGroupProcessor buildMessageWithListPayload = messageGroup -> MessageBuilder.withPayload(messageGroup.streamMessages()
-        .map(Message::getPayload)
-        .collect(Collectors.toList()))
-      .build();
+            .map(Message::getPayload)
+            .collect(Collectors.toList()))
+        .build();
     private final ReleaseStrategy listSizeReached = r -> r.size() == r.getSequenceSize();
 
-    public IntegrationConfiguration(MessageChannel receiveTextChannel, MessageChannel splitWordsChannel, MessageChannel toLowerCaseChannel, MessageChannel countWordsChannel, MessageChannel returnResponseChannel) {
+    public IntegrationConfiguration(@Qualifier("receiveTextChannel") MessageChannel receiveTextChannel,
+        @Qualifier("splitWordsChannel") MessageChannel splitWordsChannel, @Qualifier("toLowerCaseChannel") MessageChannel toLowerCaseChannel,
+        @Qualifier("countWordsChannel") MessageChannel countWordsChannel, @Qualifier("returnResponseChannel") MessageChannel returnResponseChannel) {
         this.receiveTextChannel = receiveTextChannel;
         this.splitWordsChannel = splitWordsChannel;
         this.toLowerCaseChannel = toLowerCaseChannel;
@@ -46,36 +48,36 @@ public class IntegrationConfiguration {
 
     @Bean
     public IntegrationFlow receiveText() {
-        return IntegrationFlows.from(receiveTextChannel)
-          .channel(splitWordsChannel)
-          .get();
+        return IntegrationFlow.from(receiveTextChannel)
+            .channel(splitWordsChannel)
+            .get();
     }
 
     @Bean
     public IntegrationFlow splitWords() {
-        return IntegrationFlows.from(splitWordsChannel)
-          .transform(splitWordsFunction)
-          .channel(toLowerCaseChannel)
-          .get();
+        return IntegrationFlow.from(splitWordsChannel)
+            .transform(splitWordsFunction)
+            .channel(toLowerCaseChannel)
+            .get();
     }
 
     @Bean
     public IntegrationFlow toLowerCase() {
-        return IntegrationFlows.from(toLowerCaseChannel)
-          .split()
-          .transform(toLowerCase)
-          .aggregate(aggregatorSpec -> aggregatorSpec.releaseStrategy(listSizeReached)
-            .outputProcessor(buildMessageWithListPayload))
-          .channel(countWordsChannel)
-          .get();
+        return IntegrationFlow.from(toLowerCaseChannel)
+            .split()
+            .transform(toLowerCase)
+            .aggregate(aggregatorSpec -> aggregatorSpec.releaseStrategy(listSizeReached)
+                .outputProcessor(buildMessageWithListPayload))
+            .channel(countWordsChannel)
+            .get();
     }
 
     @Bean
     public IntegrationFlow countWords() {
-        return IntegrationFlows.from(countWordsChannel)
-          .transform(convertArrayListToCountMap)
-          .channel(returnResponseChannel)
-          .get();
+        return IntegrationFlow.from(countWordsChannel)
+            .transform(convertArrayListToCountMap)
+            .channel(returnResponseChannel)
+            .get();
     }
 
 }
