@@ -1,5 +1,6 @@
 package com.baeldung.seda.springintegration;
 
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.task.TaskExecutor;
@@ -15,7 +16,11 @@ public class ChannelConfiguration {
     private final TaskExecutor countWordsChannelThreadPool;
     private final TaskExecutor returnResponseChannelThreadPool;
 
-    public ChannelConfiguration(TaskExecutor receiveTextChannelThreadPool, TaskExecutor splitWordsChannelThreadPool, TaskExecutor toLowerCaseChannelThreadPool, TaskExecutor countWordsChannelThreadPool, TaskExecutor returnResponseChannelThreadPool) {
+    public ChannelConfiguration(@Qualifier("receiveTextChannelThreadPool") TaskExecutor receiveTextChannelThreadPool,
+        @Qualifier("splitWordsChannelThreadPool") TaskExecutor splitWordsChannelThreadPool,
+        @Qualifier("toLowerCaseChannelThreadPool") TaskExecutor toLowerCaseChannelThreadPool,
+        @Qualifier("countWordsChannelThreadPool") TaskExecutor countWordsChannelThreadPool,
+        @Qualifier("returnResponseChannelThreadPool") TaskExecutor returnResponseChannelThreadPool) {
         this.receiveTextChannelThreadPool = receiveTextChannelThreadPool;
         this.splitWordsChannelThreadPool = splitWordsChannelThreadPool;
         this.toLowerCaseChannelThreadPool = toLowerCaseChannelThreadPool;
@@ -26,31 +31,31 @@ public class ChannelConfiguration {
     @Bean(name = "receiveTextChannel")
     public MessageChannel getReceiveTextChannel() {
         return MessageChannels.executor("receive-text", receiveTextChannelThreadPool)
-          .get();
+            .getObject();
     }
 
     @Bean(name = "splitWordsChannel")
     public MessageChannel getSplitWordsChannel() {
         return MessageChannels.executor("split-words", splitWordsChannelThreadPool)
-          .get();
+            .getObject();
     }
 
     @Bean(name = "toLowerCaseChannel")
     public MessageChannel getToLowerCaseChannel() {
         return MessageChannels.executor("to-lower-case", toLowerCaseChannelThreadPool)
-          .get();
+            .getObject();
     }
 
     @Bean(name = "countWordsChannel")
     public MessageChannel getCountWordsChannel() {
         return MessageChannels.executor("count-words", countWordsChannelThreadPool)
-          .get();
+            .getObject();
     }
 
     @Bean(name = "returnResponseChannel")
     public MessageChannel getReturnResponseChannel() {
         return MessageChannels.executor("return-response", returnResponseChannelThreadPool)
-          .get();
+            .getObject();
     }
 
 }
